requirements.txt
flask==3.0.0
pydantic-ai==0.1.0
litellm==1.16.0
tavily-python==0.1.9
python-dotenv==1.0.0
nest-asyncio==1.5.8